requirements.txt
coloredlogs==15.0.1
fastavro==1.8.3
pandas==2.2.0
PyYAML==6.0.1
streamlit==1.30.0
